html-speaker: Add audio player, migrate websocket.js to websocket.ts, implement audio playout
