+++ html-speaker/Makefile
@@ -1,0 +1,5 @@
+.PHONY: build
+
+build:
+	node esbuild.js
+	cp src/index.html dist/index.html
