+++ html-speaker/src/index.html
@@ -10,8 +10,9 @@
 <main>
     <h3>Messages</h3>
     <div id="messages"></div>
+<!--    <audio id="audio-player" autoplay></audio>-->
 </main>
-<script type="text/javascript" src="scripts/websocket.js"></script>
 <script type="module" src="scripts/index.js"></script>
+<!--<script src="audio.js"></script>-->
 </body>
 </html>
